+++ R/utils.R
@@ -1,25 +1,31 @@
 #' Perform Tukey Test and calculate letter placement
-#' @importFrom stats TukeyHSD aov as.formula median quantile reorder
+#' @importFrom stats TukeyHSD aov as.formula median quantile reformulate reorder
 #' @importFrom dplyr left_join
 #' @importFrom rlang as_name
 #' @param data A data.frame in long format
 #' @param x Independent variable or vector of variables to plot on x axis
 #' @param y Response variable to plot on y axis
 #' @param test Which test to run for pairwise comparisons. Either \code{tukey}
-#' (\code{\link[stats]{TukeyHSD}}) or \code{\link[pgirmess]{kruskalmc}}. Defaults
+#' (\code{\link[stats]{TukeyHSD}}), \code{\link[pgirmess]{kruskalmc}}, or
+#' \code{\link[rstatix]{dunn_test}}. Defaults
 #' to \code{tukey}.
 #' @param where Where to put the letters. Either above the box (\code{box}),
 #' above the upper whisker (\code{whisker}), or at the \code{mean} or
 #' \code{median}.
 #' @param threshold Statistical threshold for significance. Defaults to 0.05.
+#' @param reversed Logical. Argument to
+#' \code{\link[multcompView]{multcompLetters3}}. Determines whether order of
+#' letters should be reversed. Defaults to \code{FALSE}.
 #' @noRd
 
-get_tukey_letters <- function(data, x, y, group = NULL, test = c("tukey", "kruskalmc"),
+get_tukey_letters <- function(data, x, y, group = NULL,
+                              test = c("tukey", "kruskalmc", "dunn"),
                               type = c("two-way", "one-way"),
                               where = c("box", "whisker", "mean",
-                                        "median", "se", "sd", "cl_normal","cl_boot"),
-                              threshold = 0.05){
-  test <- match.arg(test, c("tukey", "kruskalmc"))
+                                        "median", "se", "sd",
+                                        "cl_normal", "cl_boot"),
+                              threshold = 0.05, reversed = FALSE){
+  test <- match.arg(test, c("tukey", "kruskalmc", "dunn"))
   where <- match.arg(where, c("box", "whisker", "mean","median", "se", "sd",
                               "cl_normal", "cl_boot"))
   type <- match.arg(type, c("two-way", "one-way"))
@@ -29,17 +35,19 @@ get_tukey_letters <- function(data, x, y, group = NULL, test = c("tukey", "krusk
   if (inherits(x, "quosure") & is.null(group)){
     letters.df <- place_tukey_letters(data = data, x = as_name(x), y = as_name(y),
                                       test = test, where = where,
-                                      threshold = threshold)
+                                      threshold = threshold, reversed = reversed)
   } else{
     if (type == "two-way"){
       letters.df <- place_tukey_letters(data = data, x = sapply(x, as_name),
                                         y = as_name(y), test = test,
-                                        where = where, threshold = threshold)
+                                        where = where, threshold = threshold,
+                                        reversed = reversed)
     } else if (type == "one-way"){
       letters.df <- purrr::map_dfr(unique(data[[as_name(group)]]), function(gr){
         data %>% filter(!!group == gr) %>%
           place_tukey_letters(x = as_name(x), y = as_name(y), test = test,
-                              where = where, threshold = threshold) %>%
+                              where = where, threshold = threshold,
+                              reversed = reversed) %>%
           mutate(!!group := gr) %>% tibble::remove_rownames()
       })
     }
@@ -49,27 +57,41 @@ get_tukey_letters <- function(data, x, y, group = NULL, test = c("tukey", "krusk
 
 #' Place Tukey Letters
 #' @noRd
-place_tukey_letters <- function(data, x, y, test = c("tukey", "kruskalmc"),
-                                where = c("box","whisker"),
-                                threshold=0.05){
+place_tukey_letters <- function(data, x, y, test = c("tukey", "kruskalmc", "dunn"),
+                                where = c("box", "whisker"),
+                                threshold = 0.05, p.adjust, reversed = FALSE){
   if (length(x) == 1){
-    form <- as.formula(paste(y, x, sep="~"))
+    form <- form_let <- reformulate(x,y)
     xlab <- x
   } else {
-    form <- as.formula(paste(y, paste(x, collapse="*"), sep="~"))
+    form <- reformulate(termlabels = paste(x, collapse="*"), y)
+    form_let <- reformulate(termlabels = paste(x, collapse=":"), y)
     xlab <- paste(x, collapse=":")
     data[,xlab] <- apply(data[,x], 1, paste, collapse = ":")
   }
   if (test == "tukey"){
-    tukey <- TukeyHSD(aov(form, data = data))[[xlab]][,4]
-    tukey <- tukey[which(!is.na(tukey))]
-    letters.df <- data.frame("Letter" = multcompLetters(tukey, threshold = threshold)$Letters)
+    tukey <- TukeyHSD(mod <- aov(form, data = data))
+    # tukey <- tukey[which(!is.na(tukey))]
+    # letters.df <- data.frame("Letter" = multcompLetters(tukey, threshold = threshold)$Letters)
+    # let <- multcompView::multcompLetters2(form_let, tukey[[xlab]][,"p adj"],
+    #                                       data = as.data.frame(data))
+    diff <- tukey[[xlab]][,"p adj"]
+    # multcompLetters4(mod, tukey)$treatments
   } else if (test == "kruskalmc"){
     test <- pgirmess::kruskalmc(form, data = data, probs = threshold)
     diff <- test$dif.com[,"stat.signif"]
     names(diff) <- rownames(test$dif.com)
-    letters.df <- data.frame("Letter" = multcompLetters(diff)$Letters)
+    # letters.df <- data.frame("Letters" = multcompLetters2(formula = diff)$Letters)
+  } else if (test == "dunn"){
+    test <- rstatix::dunn_test(form_let, data = data)
+    diff <- test$p
+    names(diff) <- paste(test$group1, test$group2, sep="-")
   }
+  let <- multcompView::multcompLetters3(z = xlab, y = y,
+                                        x = diff,
+                                        data = as.data.frame(data),
+                                        reversed = reversed)
+  letters.df <- data.frame(Letter = let$Letters)
   # letters.df <- data.frame("Letter" = multcompLetters(tukey, threshold = threshold)$Letters)
   letters.df[[xlab]] <- rownames(letters.df) #Create column based on rownames
 
@@ -83,11 +105,11 @@ place_tukey_letters <- function(data, x, y, test = c("tukey", "kruskalmc"),
                           "cl_normal" = get_cl_normal,
                           "cl_boot" = get_cl_boot)
   placement <- data %>% # Create a dataframe to assign the letter position.
-    group_by(.data[[xlab]]) %>%
-    summarise("Placement.Value" = placement_fnc(.data[[y]]))
-  letters.df <- left_join(letters.df, placement, by = xlab) # Merge dataframes
+    dplyr::group_by(.data[[xlab]]) %>%
+    dplyr::summarise("Placement.Value" = placement_fnc(.data[[y]]))
+  letters.df <- dplyr::left_join(letters.df, placement, by = xlab) # Merge dataframes
   if (length(x) > 1){
-    letters.df <- left_join(letters.df, unique(data[,c(xlab, x)]), by = xlab)
+    letters.df <- dplyr::left_join(letters.df, unique(data[,c(xlab, x)]), by = xlab)
   }
   letters.df
 }
@@ -178,14 +200,14 @@ is.color <- function(x, return.colors = FALSE) {
 #' @importFrom stats quantile
 #' @noRd
 get_quantile <- function(x){
-  quantile(x, na.rm=TRUE)[4]
+  quantile(x, na.rm = TRUE)[4]
 }
 
 #' Calculate boxplot whisker
 #' @importFrom stats IQR
 #'@noRd
 get_whisker <- function(x){
-  r <- quantile(x, na.rm=TRUE)[4] + 1.5*IQR(x, na.rm=TRUE)
-  x <- x[x<=r]
-  max(x, na.rm=TRUE)
+  r <- quantile(x, na.rm = TRUE)[4] + 1.5*IQR(x, na.rm = TRUE)
+  x <- x[x <= r]
+  max(x, na.rm = TRUE)
 }